Refactor book routes into flask blueprint and add code into the correct files
